requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
json5==0.10.0
packaging==24.2
playwright==1.49.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
